allow time to go backwards

+++ NTPClock.cpp
@@ -19,11 +19,11 @@ uint8_t NTPClock::getTime(uint32_t now, uint32_t *ntpTimestamp, uint32_t *ntpFra
   if (!timeset_)
     return 0;
 
-  uint64_t ntpFracPassed = (now - lastMicros_) * 4294967296ULL / 1000000ULL;
-  int32_t ntpFracPassedDrift = (int64_t)ntpFracPassed * ppb_ / 1000000000LL;
+  int64_t ntpFracPassed = (now - lastMicros_) * 4294967296LL / 1000000LL;
+  int32_t ntpFracPassedDrift = ntpFracPassed * ppb_ / 1000000000LL;
   ntpFracPassed += ntpFracPassedDrift;
   temp_.whole = ntpTimestamp_.whole + ntpFracPassed;
-  if(ntpFracPassed >= 4294967296ULL) { // every second
+  if(ntpFracPassed >= 4294967296LL) { // every second
     lastMicros_ = now;
     ntpTimestamp_.whole = temp_.whole;
   }
